\name{dbExample}
\docType{data}
\alias{dbExample}
\title{Simulated database with 1,000 individuals}
\description{
  Database containing 1,000 simulated DNA profiles typed on ten autosomal markers.
}
\usage{dbExample}
\format{A data frame with each row being a DNA profile and each column a
part of a genetic marker. Note that homozygote profiles has the same
allelic value in the two columns associated to the same marker.}
