% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNAtools-package.R, R/zzz.R
\docType{package}
\name{DNAtools}
\alias{DNAtools}
\alias{DNAtools-package}
\title{DNAtools}
\description{
Tools for Analysing Forensic Genetic DNA Data

Tools for Analysing Forensic Genetic DNA Data
}
\author{
Torben Tvedebrink <tvede@math.aau.dk> and James Curran
<j.curran@auckland.ac.nz>. Mikkel Meyer Andersen <mikl@math.aau.dk>
assisted on the multicore/CPU implementation.

Torben Tvedebrink <tvede@math.aau.dk>, James Curran <j.curran@auckland.ac.nz>, and Mikkel Meyer Andersen <mikl@math.aau.dk>.
}
